Add afs-install-tarball-client
This script installs a version of the tarball osg-client into /p/vdt/workspaces
on AFS. The version to install is passed on the command line. It sets up
symlinks for the AFS @sys magic, runs osg-post-install, and gets CA certs.

The idea is for this script to be run as part of the release process, so that
we would have a way of running grid tools on UW CSL-supported machines.


git-svn-id: https://vdt.cs.wisc.edu/svn/software/tools@17301 4e558342-562e-0410-864c-e07659590f8c