tool for finding package version differences between osg tags (SOFTWARE-2073)
This is basically 'osg-outdated-epel-pkgs' stripped down to work with
comparing two osg tags instead of osg vs epel.  But it works the same way as
the old script in that it finds the package overlap between tag1 and tag2,
and reports the builds that have newer versions in tag2 than tag1.

This is mostly intended to be used to identify packages with newer builds in
development than in testing, in case we forgot to promote a build that was
planned for a release.  ... But it can also be used, for example, to compare
versions between two different osg release series.


git-svn-id: https://vdt.cs.wisc.edu/svn/software/tools@21820 4e558342-562e-0410-864c-e07659590f8c