scripts for listing what package versions builds were built against
the 'built-against-pkgs' one was requested by TimC.

'list-buildroot' is more general purpose and gets used by the other.

An example run for the jira table TimC was asking about:

  ./built-against-pkgs -t htcondor-ce-1.19-1.osg{32,33}.el{5,6,7} -s condor

Run with no arguments (or --help) for usage.

(Sorry, there doesn't seem to be an osg-koji way to list all builds of a
package, apart from tagging information, so listing the builds to query
is necessary -- or at least a superset of the valid ones.)

git-svn-id: https://vdt.cs.wisc.edu/svn/software/tools@22069 4e558342-562e-0410-864c-e07659590f8c