handle case where build NVR, arch A, has no corresponding NVR.A rpm
In particular, this handles the bestman2 noarch builds, which have
bestman2-client-V-R.noarch rpms, but not bestman2-V-R.noarch.

So, we now make an important distinction between specifying "NVR arch",
which is the build NVR, with target arch, and "NVR.arch", which any binary
rpm's full NVRA, not necessarily matching the (source) package name of the
build itself.


git-svn-id: https://vdt.cs.wisc.edu/svn/software/tools@22091 4e558342-562e-0410-864c-e07659590f8c