quota notification script (SOFTWARE-1666)
The script `afs-quota-notify' is intended to run as a cron job, to
report on the quota usage of various directories in AFS. The script
operates on directories, not volumes, which may lead to some duplicate
warnings if multiple directories are mounted on the same AFS volume.
Still, IMO, reporting on directories makes more sense to the user.

I included a config file called `afs-quota-notify-vdt.ini' for
monitoring the AFS dirs that OSG cares about and sending the OSG Madison
people email, but I expect this script might be useful for more than one
group.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/tools@22593 4e558342-562e-0410-864c-e07659590f8c