fix unicode handling of stdin for python3
in python3, sys.stdin does not get opened automatically with the utf-8
encoding option, so stdin (fd 0) needs to be fdopened.  Otherwise
a json file with non-ascii unicode characters can cause an exception
when loaded by the json module.

Note in python3 open behaves like os.fdopen if an int is passed in for
the path, but in python2 we have to use the right one.

Also, for '-', make the default dest 'stdin%' rather than '-%', to
avoid creating a path with a leading hyphen,