fix .ts file creation
there was a chicken-and-the-egg problem here.

want_since was true in get_since_kw if the itemgetter took a since arg.

but it only returned a since parameter if the .ts file existed.

and the .ts file was only written (want_since for dump_items) if
the since parameter was returned.

so at some point apparently we created this loop where the existing
.ts files would be used but no new ones would ever be created.

Anyway, move the want_since calculation out of get_since_kw into the
calling dump_updated_obj_items, and pass that into dump_items, rather
than using "'since' in kw", which had the additional condition that
the .ts file already existed.