improve el8 detection regex heuristic
sometimes the %dist for el8 packages contains a '.el8_' rather than
a '.el8.', and if that happens for the final package in the install
list, it will guess the wrong log format.  This little tweak has solved
the issue for me consistently for the last few months at least.

Cheers.