Add svn-to-monorepo and helper script
This converts the VDT SVN repo to a Git repo with a 'monorepo' layout
-- meaning all of the branches are in a directory, instead of there
being one Git branch per SVN branch.  This will be used for creating
and updating the https://github.com/osg-htc/software-packaging repo.

The `git filter-repo` breaks the link to SVN so you will need to
rerun the script to do a full re-conversion in order to get new SVN
commits (as opposed to just running `git svn rebase` and pushing),
but subsequent conversions should still be fast-forward so no force
pushing should be necessary.